+++ R/betafunctions.R
@@ -9,10 +9,10 @@
 #' @param orders The number of moment-orders to be calculated for each of the moment-types.
 #' @examples
 #' # Assume some variable follows a four-parameter Beta distribution with
-#' # location parameters l = 0.25 and u = 0.75, and shape parameters a = 5
-#' # and b = 3. To compute the first four raw, central, and standardized
+#' # location parameters l = 0.25 and u = 0.75, and shape parameters alpha = 5
+#' # and beta = 3. To compute the first four raw, central, and standardized
 #' # moments of this distribution using betamoments():
-#' betamoments(a = 5, b = 3, l = 0.25, u = 0.75,
+#' betamoments(alpha = 5, beta = 3, l = 0.25, u = 0.75,
 #' types = c("raw", "central", "standardized"), orders = 4)
 #' @references Hanson, B. A (1991). Method of Moments Estimates for the Four-Parameter Beta Compound Binomial Model and the Calculation of Classification Consistency Indexes. American College Testing Research Report Series.
 #' @return A list of moment types, each a list of moment orders.